Split 'test_routers_network_ops.py' into 'test_isolated_networks.py' and 'test_redundant_isolated_networks.py'
